+++ src/storage-blob-preview/azext_storage_blob_preview/vendored_sdks/azure_storage_blob/v2020_02_10/_generated/operations/_append_blob_operations.py
@@ -65,25 +65,25 @@ class AppendBlobOperations(object):
          blob operations.
         :type blob_tags_string: str
         :param blob_http_headers: Additional parameters for the operation
-        :type blob_http_headers: ~azure.storage.blob.models.BlobHTTPHeaders
+        :type blob_http_headers: ~.models.BlobHTTPHeaders
         :param lease_access_conditions: Additional parameters for the
          operation
         :type lease_access_conditions:
-         ~azure.storage.blob.models.LeaseAccessConditions
+         ~.models.LeaseAccessConditions
         :param cpk_info: Additional parameters for the operation
-        :type cpk_info: ~azure.storage.blob.models.CpkInfo
+        :type cpk_info: ~.models.CpkInfo
         :param cpk_scope_info: Additional parameters for the operation
-        :type cpk_scope_info: ~azure.storage.blob.models.CpkScopeInfo
+        :type cpk_scope_info: ~.models.CpkScopeInfo
         :param modified_access_conditions: Additional parameters for the
          operation
         :type modified_access_conditions:
-         ~azure.storage.blob.models.ModifiedAccessConditions
+         ~.models.ModifiedAccessConditions
         :param callable cls: A custom type or function that will be passed the
          direct response
         :return: None or the result of cls(response)
         :rtype: None
         :raises:
-         :class:`StorageErrorException<azure.storage.blob.models.StorageErrorException>`
+         :class:`StorageErrorException<.models.StorageErrorException>`
         """
         error_map = kwargs.pop('error_map', None)
         blob_content_type = None
@@ -246,25 +246,25 @@ class AppendBlobOperations(object):
         :param lease_access_conditions: Additional parameters for the
          operation
         :type lease_access_conditions:
-         ~azure.storage.blob.models.LeaseAccessConditions
+         ~.models.LeaseAccessConditions
         :param append_position_access_conditions: Additional parameters for
          the operation
         :type append_position_access_conditions:
-         ~azure.storage.blob.models.AppendPositionAccessConditions
+         ~.models.AppendPositionAccessConditions
         :param cpk_info: Additional parameters for the operation
-        :type cpk_info: ~azure.storage.blob.models.CpkInfo
+        :type cpk_info: ~.models.CpkInfo
         :param cpk_scope_info: Additional parameters for the operation
-        :type cpk_scope_info: ~azure.storage.blob.models.CpkScopeInfo
+        :type cpk_scope_info: ~.models.CpkScopeInfo
         :param modified_access_conditions: Additional parameters for the
          operation
         :type modified_access_conditions:
-         ~azure.storage.blob.models.ModifiedAccessConditions
+         ~.models.ModifiedAccessConditions
         :param callable cls: A custom type or function that will be passed the
          direct response
         :return: None or the result of cls(response)
         :rtype: None
         :raises:
-         :class:`StorageErrorException<azure.storage.blob.models.StorageErrorException>`
+         :class:`StorageErrorException<.models.StorageErrorException>`
         """
         error_map = kwargs.pop('error_map', None)
         lease_id = None
@@ -418,31 +418,31 @@ class AppendBlobOperations(object):
          analytics logging is enabled.
         :type request_id: str
         :param cpk_info: Additional parameters for the operation
-        :type cpk_info: ~azure.storage.blob.models.CpkInfo
+        :type cpk_info: ~.models.CpkInfo
         :param cpk_scope_info: Additional parameters for the operation
-        :type cpk_scope_info: ~azure.storage.blob.models.CpkScopeInfo
+        :type cpk_scope_info: ~.models.CpkScopeInfo
         :param lease_access_conditions: Additional parameters for the
          operation
         :type lease_access_conditions:
-         ~azure.storage.blob.models.LeaseAccessConditions
+         ~.models.LeaseAccessConditions
         :param append_position_access_conditions: Additional parameters for
          the operation
         :type append_position_access_conditions:
-         ~azure.storage.blob.models.AppendPositionAccessConditions
+         ~.models.AppendPositionAccessConditions
         :param modified_access_conditions: Additional parameters for the
          operation
         :type modified_access_conditions:
-         ~azure.storage.blob.models.ModifiedAccessConditions
+         ~.models.ModifiedAccessConditions
         :param source_modified_access_conditions: Additional parameters for
          the operation
         :type source_modified_access_conditions:
-         ~azure.storage.blob.models.SourceModifiedAccessConditions
+         ~.models.SourceModifiedAccessConditions
         :param callable cls: A custom type or function that will be passed the
          direct response
         :return: None or the result of cls(response)
         :rtype: None
         :raises:
-         :class:`StorageErrorException<azure.storage.blob.models.StorageErrorException>`
+         :class:`StorageErrorException<.models.StorageErrorException>`
         """
         error_map = kwargs.pop('error_map', None)
         encryption_key = None
@@ -601,21 +601,21 @@ class AppendBlobOperations(object):
         :param lease_access_conditions: Additional parameters for the
          operation
         :type lease_access_conditions:
-         ~azure.storage.blob.models.LeaseAccessConditions
+         ~.models.LeaseAccessConditions
         :param modified_access_conditions: Additional parameters for the
          operation
         :type modified_access_conditions:
-         ~azure.storage.blob.models.ModifiedAccessConditions
+         ~.models.ModifiedAccessConditions
         :param append_position_access_conditions: Additional parameters for
          the operation
         :type append_position_access_conditions:
-         ~azure.storage.blob.models.AppendPositionAccessConditions
+         ~.models.AppendPositionAccessConditions
         :param callable cls: A custom type or function that will be passed the
          direct response
         :return: None or the result of cls(response)
         :rtype: None
         :raises:
-         :class:`StorageErrorException<azure.storage.blob.models.StorageErrorException>`
+         :class:`StorageErrorException<.models.StorageErrorException>`
         """
         error_map = kwargs.pop('error_map', None)
         lease_id = None