+++ src/aks-preview/HISTORY.md
@@ -5,6 +5,7 @@ Release History
 0.4.72
 ++++++
 * Add --no-uptime-sla
+* Create MSI clusters by default.
 
 0.4.71
 ++++++