+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,15 @@
 
 Release History
 ===============
+0.4.72
+++++++
+* Add --no-uptime-sla
+* Create MSI clusters by default.
+
+0.4.71
+++++++
+* Add support using custom private dns zone resource id for parameter '--private-dns-zone'
+
 0.4.70
 ++++++
 * Revert to use CLIError to be compatible with azure cli versions < 2.15.0